PiperOrigin-RevId: 337132590

+++ tensorflow_probability/python/layers/distribution_layer.py
@@ -92,7 +92,7 @@ def _event_size(event_shape, name=None):
 
     event_shape_const = tf.get_static_value(event_shape)
     if event_shape_const is not None:
-      return np.prod(event_shape_const)
+      return np.prod(event_shape_const, dtype=np.int32)
     else:
       return tf.reduce_prod(event_shape)
 
@@ -896,7 +896,7 @@ class IndependentLogistic(DistributionLambda):
     with tf.name_scope(name or 'IndependentLogistic_params_size'):
       event_shape = tf.convert_to_tensor(
           event_shape, name='event_shape', dtype_hint=tf.int32)
-      return 2 * _event_size(
+      return np.int32(2) * _event_size(
           event_shape, name=name or 'IndependentLogistic_params_size')
 
   def get_config(self):
@@ -1011,7 +1011,7 @@ class IndependentNormal(DistributionLambda):
     with tf.name_scope(name or 'IndependentNormal_params_size'):
       event_shape = tf.convert_to_tensor(
           event_shape, name='event_shape', dtype_hint=tf.int32)
-      return 2 * _event_size(
+      return np.int32(2) * _event_size(
           event_shape, name=name or 'IndependentNormal_params_size')
 
   def get_config(self):
